make all deploy args configurable
