+++ types/utils.ts
@@ -111,3 +111,28 @@ export function readBuffer(slice: Slice) {
 
     return res;
 }
+
+// returns a string with a decimal point
+export function formatCoin(coins: bigint): string {
+    const divisor = 1_000_000_000n;
+    const tons = coins / divisor;
+    const fractional = coins % divisor;
+
+    if (fractional === 0n) {
+        return String(tons);
+    }
+
+    // Ensure the fractional part is always 9 digits by padding with leading zeros if necessary
+    let fractionalStr = fractional.toString().padStart(9, '0');
+    fractionalStr = trimSuffix(fractionalStr, '0');
+
+    return `${tons.toString()}.${fractionalStr}`;
+}
+
+function trimSuffix(str: string, suffix: string): string {
+    while (str.endsWith(suffix)) {
+        str = str.slice(0, -suffix.length);
+    }
+
+    return str;
+}
